production-pipeline/requirements_db.txt
# Database Integration Requirements for Trivia Video Pipeline V2.0

# Core Flask and Web Framework
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1

# Google Cloud Services
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-texttospeech==2.16.3
google-generativeai==0.3.2

# Video Processing
ffmpeg-python==0.2.0

# Database and Data Processing
pandas==2.1.1
numpy==1.25.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1

# Development and Testing
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0

# Async Support
aiohttp==3.8.6
asyncio-throttle==1.0.2

# Logging and Monitoring
structlog==23.1.0
prometheus-client==0.17.1

# Security
cryptography==41.0.4

